refactor: move yaml-specific code to separate file
